Updated to API and bug fixes
+ Put API into controller and routes
+ Add cursor to Create Post and Edit Post
