+++ api/index.js
@@ -1,167 +1,36 @@
 const express = require("express");
 const cors = require("cors");
+
+// Database requirements
 const { default: mongoose } = require("mongoose");
-const bcrypt = require("bcryptjs");
-const jwt = require("jsonwebtoken");
-const multer = require("multer");
-const fs = require("fs");
+const { ObjectId } = require("mongodb");
 
 const cookieParser = require("cookie-parser");
-const User = require("./models/User");
-const Post = require("./models/Post");
-const { ObjectId } = require("mongodb");
+
+// Routes requirements
+const userRoutes = require("./routes/userRoutes");
+const postRoutes = require("./routes/postRoutes");
+
 const app = express();
-const uploadMiddleware = multer({ dest: "uploads/" });
 
-const passwordSalt = bcrypt.genSaltSync(10);
-const tokenSecret = "bcrypt.genSaltSync(10)";
-
+// Use cors
 app.use(cors({ credentials: true, origin: "http://localhost:3000" }));
+
+// Middleware to parse incoming JSON data
 app.use(express.json());
+
+// Middleware to parse cookies
 app.use(cookieParser());
+
+// Upload path
 app.use("/uploads", express.static(__dirname + "/uploads"));
+
+// Connect to MongoDB
 mongoose.connect(
   "mongodb+srv://phamtruc-work:phamtruc-work@cluster0.nhzjq83.mongodb.net/?retryWrites=true&w=majority"
 );
 
-app.post("/register", async (req, res) => {
-  const { username, password } = req.body;
-  try {
-    const UserDoc = await User.create({
-      username,
-      password: bcrypt.hashSync(password, passwordSalt),
-    });
-    res.json(UserDoc);
-  } catch (e) {
-    res.status(400).json(e);
-  }
-});
-
-app.post("/login", async (req, res) => {
-  const { username, password } = req.body;
-  if (!username || !password) {
-    return res.status(401).json("Username or Password is null");
-  }
-  const UserDoc = await User.findOne({ username });
-
-  const passwordCheck = bcrypt.compareSync(password, UserDoc.password);
-
-  if (passwordCheck) {
-    jwt.sign({ username, id: UserDoc._id }, tokenSecret, {}, (err, token) => {
-      if (err) throw err;
-      res.cookie("token", token).json({ id: UserDoc._id, username });
-      res.status(200);
-    });
-  } else {
-    res.status(400).json("Wrong Credentials");
-  }
-});
-
-app.get("/profile", (req, res) => {
-  const { token } = req.cookies;
-  if (token) {
-    jwt.verify(token, tokenSecret, {}, (err, info) => {
-      if (err) throw err;
-      res.json(info);
-    });
-  }
-});
-
-app.post("/logout", (req, res) => {
-  res.cookie("token", "").json("ok");
-});
-
-app.post("/post", uploadMiddleware.single("file"), async (req, res) => {
-  if (!req.file) {
-    return res.status(401).json("You need to choose a cover");
-  }
-  const { originalname, path } = req.file;
-  const parts = originalname.split(".");
-  const ext = parts[parts.length - 1];
-  const newPath = path + "." + ext;
-  fs.renameSync(path, newPath);
-
-  const { token } = req.cookies;
-
-  if (token) {
-    jwt.verify(token, tokenSecret, {}, async (err, info) => {
-      if (err) throw err;
-      const { title, summary, content } = req.body;
-      if (!title || !summary || !content) {
-        return res
-          .status(400)
-          .json("You need to write a title/summary/content");
-      }
-      const postDoc = await Post.create({
-        title,
-        summary,
-        content,
-        cover: newPath,
-        author: info.id,
-      });
-      res.status(200).json(postDoc);
-    });
-  }
-});
-
-app.put("/post", uploadMiddleware.single("file"), async (req, res) => {
-  const newPath = null;
-  if (req.file) {
-    const { originalname, path } = req.file;
-    const parts = originalname.split(".");
-    const ext = parts[parts.length - 1];
-    newPath = path + "." + ext;
-    fs.renameSync(path, newPath);
-  }
-
-  const { token } = req.cookies;
-
-  jwt.verify(token, tokenSecret, {}, async (err, info) => {
-    if (err) throw err;
-    const { id, title, summary, content } = req.body;
-
-    const postDoc = await Post.findById(id);
-    const isAuthor = JSON.stringify(postDoc.author) === JSON.stringify(info.id);
-    if (!isAuthor) {
-      res.status(400).json("You are not the author");
-    }
-    await postDoc.updateOne({
-      title,
-      summary,
-      content,
-      cover: newPath ? newPath : postDoc.cover,
-    });
-    res.status(200).json(postDoc);
-  });
-});
-
-app.delete("/post/:id", async (req, res) => {
-  if (ObjectId.isValid(req.params.id)) {
-    Post.deleteOne({ _id: Object(req.params.id) })
-      .then((result) => {
-        res.status(200).json(result);
-      })
-      .catch((err) => {
-        res.status(500).json({ error: "Could not delete the document" });
-      });
-  } else {
-    res.status(500).json({ error: "Could not find the document" });
-  }
-});
-
-app.get("/post", async (req, res) => {
-  res.json(
-    await Post.find()
-      .populate("author", ["username"])
-      .sort({ createdAt: -1 })
-      .limit(20)
-  );
-});
-
-app.get("/post/:id", async (req, res) => {
-  const { id } = req.params;
-  const postDoc = await Post.findById(id).populate("author", ["username"]);
-  res.json(postDoc);
-});
+app.use(userRoutes);
+app.use(postRoutes);
 
 app.listen(4000);
